fix: move command-based bool ops to geometry commands
